assigned random up down movement to jellyFish.class.js
