+++ js/classes/moveable-object.class.js
@@ -1,5 +1,3 @@
-import { moveObjRatio } from "../script.js";
-
 export class MoveableObject {
   img;
   x = 100;
